Rename: api/login.ts -> auth.ts
